Implement incremental (fast) softmax.
Computes softmax with two loads to memory instead of three. In theory, this
should be faster since reading from memory is slow. This might be the case for GPUs
but some naive benchmarks I tried show the "faster" version is actually slower on CPU.
This is very likely because softmax memory access pattern is very straightforward and
the CPU can predict which memory chunks will be needed and frontrun the CPU. So the
actual runtime is dominanted by the extra compute of the "faster" version.